automaton: add new accept_eof method
This is to support automata that delay matches by one byte. (Which is
what regex-automata will do in its 0.2 release, in order to support
look-around assertions such as ^, $ and \b.)

This requires a small tweak to the search code to call this new method
once it sees a final node.